requirements.txt
crawl4ai
python-dotenv
nemoguardrails[openai]
llama-index
llama-index-vector-stores-supabase
llama-index-embeddings-openai
llama-index-llms-openai
openai
streamlit
pytest
pytest-asyncio
